% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_config_CDMTools.R
\name{configCDMTools}
\alias{configCDMTools}
\title{configCDMTools}
\usage{
configCDMTools()
}
\value{
returns a cdm_webapi_conn list with the info set in get_golem_config
}
\description{
buils cdm_webapi_conn based on the info un golem yalm
}
