% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_cohort_atlas.R
\name{mod_import_cohort_atlas_server}
\alias{mod_import_cohort_atlas_server}
\title{import_cohort_atlas Server Functions}
\usage{
mod_import_cohort_atlas_server(id, r_connection, r_cohorts)
}
\description{
import_cohort_atlas Server Functions
}
