% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spiners.R
\name{sweetAlert_spinner}
\alias{sweetAlert_spinner}
\title{sweetAlert_spinner}
\usage{
sweetAlert_spinner(message, wait_time_sec = NULL, ...)
}
\arguments{
\item{message}{message to show in the alert}

\item{wait_time_sec}{if not NULL shows a progress bar that loads in the given second}

\item{...}{additional options for shinyWidgets::show_alert}
}
\value{

}
\description{
A sweet alert with a dna spinner
}
