% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spiners.R
\name{ui_load_spinner}
\alias{ui_load_spinner}
\title{ui_load_spinner}
\usage{
ui_load_spinner(ui_element, ...)
}
\arguments{
\item{ui_element}{ui_element to add spinner to}

\item{...}{other parameters to shinycustomloader::withLoader}
}
\value{

}
\description{
Adds dna spinner to a ui_element
}
